Create README.md
The aim of this case study is to learn how to deal with complex datasets and an overwhelming number of variables, and write professional scripts in R or Python. Data cleansing and feature selection or reduction methods should be used before applying data mining methods to predict defaults.

The SECOM data set in the Machine Learning Repository (2008) provides information of the manufacturing process in the semiconductor industry. In this sector, the production line is being monitored using many sensors. That’s because the production is difficult and defaults in later stages of the production are particularly expensive.

Using many sensors for monitoring the production lines results in datasets with many variables. But statistical models used to make predictions about the state of certain products should be built based on only as many variables as necessary. Such models would be parsimonious and have great exploratory power.